Rename trees utility to graphs.
